% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bin2ind}
\alias{bin2ind}
\title{convert binary codes to decimal integers}
\usage{
bin2ind(mat, LOG = FALSE)
}
\arguments{
\item{mat}{A binary matrix; Each row is a binary representation to be converted.}

\item{LOG}{Take the logarithm of the decimal integer (default is \code{TRUE});
otherwise, set to \code{FALSE}.}
}
\value{
A vector of numbers after conversion (not a matrix)
}
\description{
convert binary codes to decimal integers
}
\examples{
bin2ind(c(1,1,1))
bin2ind(rbind(c(1,0,0),c(1,1,0),c(0,0,1),c(1,0,1),c(0,0,0)))
bin2ind(rbind(c(1,0,0),c(1,1,0),c(0,0,1),c(1,0,1),c(0,0,0)),LOG=FALSE)
}
