% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{binary}
\alias{binary}
\title{convert decimal integers to binary codes}
\usage{
binary(x, noBits)
}
\arguments{
\item{x}{a vector of base 10 integers}
}
\value{
A matrix of binary forms of x (by row)
}
\description{
convert decimal integers to binary codes
}
\examples{
decimals <- c(3,5,11,4)
m <- binary(decimals)
m
}
