% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bsxfun_7_hinge_pow}
\alias{bsxfun_7_hinge_pow}
\title{7 shaped binary array operation by bsxfun from matlab but
with "hinge" shape}
\usage{
bsxfun_7_hinge_pow(A, B)
}
\arguments{
\item{A}{matrix of size d1 by d2}

\item{B}{matrix of size d3 by d2}
}
\value{
a matrix of dimension d3 by d1 by d2
}
\description{
7 shaped binary array operation by bsxfun from matlab but
with "hinge" shape
}
\examples{
bsxfun_7_hinge_pow(matrix(c(1,2,3,4,5,6),nrow=3,byrow=TRUE),
matrix(c(1:18),nrow=9,byrow=TRUE))
}
