% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_complete}
\alias{check_complete}
\title{check if matrix Q is complete}
\usage{
check_complete(Q)
}
\arguments{
\item{Q}{binary matrix}
}
\value{
a list of information
\itemize{
\item is_complete: 0 for not complete
\item ind_I_K: the single item index for each attribute;
if an attribute has multiple single-attribute items, just pick the one with
the lowest item index.
\item sum_single: number of items requiring each attribute
}
}
\description{
Check if there is an item solely requiring attribute k
for each attribute k = 1, 2, ..., K
}
\examples{
check_complete(matrix(c(1,0,0, 0, 1,0, 0, 0, 1, 1, 1, 0, 0, 0,1),
ncol=3,byrow=TRUE))
}
