% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_overlap}
\alias{check_overlap}
\title{check if patterns in A1 match those in A2}
\usage{
check_overlap(A1, A2)
}
\arguments{
\item{A1}{a binary matrix}

\item{A2}{a binary matrix with the same dimensions as \code{A1}}
}
\value{
a list of three variables
\itemize{
\item among1 a binary vector of each row of A1 identical to some row(s) of A2
\item among2 same as among1 but switch A1 and A2.
\item overlap a binary vector
}
}
\description{
check if patterns in A1 match those in A2
}
\examples{

check_overlap(matrix(c(1,1,0, 1,0,0),nrow=2,byrow=TRUE),
matrix(c(1,0,1, 1,1,0),nrow=2,byrow=TRUE))
}
