% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{f}
\alias{f}
\title{flip the matrix to the right form for \code{image()}}
\usage{
f(m)
}
\arguments{
\item{m}{a matrix}
}
\value{
a matrix after image() will look the same as the matrix
}
\description{
This function makes the result of \code{image()} to be plotted as shown in a matrix
}
