% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{generate_X_fromA}
\alias{generate_X_fromA}
\title{generate simulated data A}
\usage{
generate_X_fromA(N, A, p, Q, c, g, model = "DINA")
}
\arguments{
\item{N}{sample size}

\item{A}{K column latent state attributes (unique rows)}

\item{p}{2^K by 1 vector of prevalence probabilities}

\item{Q}{M by K item-attribute relationship}

\item{c}{M by 1 one minus slipping parameters}

\item{g}{M by 1 guessing parameters}

\item{model}{"DINA"(default) or "DINO"}
}
\value{
\itemize{
\item X: N by M matrix of responses
\item Z: N by K matrix of attributes for N individuals
\item X_true: N by M ideal response matrix
}
}
\description{
Default to DINA
}
\details{
shorter p compared to \code{\link{generate_X_DINA}}
}
\examples{
N <- 1000
A <- rbind(c(1,1,1),c(0,0,1))
p <- c(0.9,0.1)
Q <- matrix(c(0,0,0, 0,0,1, 0,1,0,
              0,1,1, 1,0,0, 1,0,1,
              1,1,0, 1,1,1),ncol=3,byrow=TRUE)
M <- nrow(Q)
c <- rep(0.8,M)
g <- rep(0.2,M)
generate_X_fromA(N,A,p,Q,c,g)
}
