% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{generate_X_tax2}
\alias{generate_X_tax2}
\title{generate second-level responses with taxonomy under DINA}
\usage{
generate_X_tax2(X1_sub, A_set20, D_mat, p_short, Q2, c, g, model = "DINA")
}
\arguments{
\item{X1_sub}{The subset of individuals who has one particular level 1 latent attributes}

\item{A_set20}{The true leve 2 set of latent attributes}

\item{D_mat}{J1 by J2 matrix for indicating the tree structure}

\item{p_short}{level 2 sparse population level prevalences}

\item{Q2}{level 2 Q in a subset of level 1 latent attributes}

\item{c}{1-slipping parameters}

\item{g}{guess parameters}

\item{model}{"DINA" (default); "DINO"}
}
\value{
\itemize{

\item X2 level 2 observed responses
\item Z latent attribute profiles for all individuals at level 2
\item X_true2 Ideal response matrix
}
}
\description{
generate second-level responses with taxonomy under DINA
}
