% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{get_EM}
\alias{get_EM}
\title{function to actually perform EM for DINA (fixed A)}
\usage{
get_EM(X, Q, A, err_prob, c = NULL, g = NULL, model = "DINA")
}
\arguments{
\item{X}{N by J data matrix}

\item{Q}{J by K structural matrix}

\item{A}{K-column latent attribute matrix}

\item{err_prob}{error probability; for initializing c g}

\item{c, g}{default to NULL, but can use values obtained from PEM.}

\item{model}{"DINA" (default) or "DINO"}
}
\value{
\itemize{
\item nu vector of potential ("Delta")
\item c 1-slipping parameter
\item g guessing parameter
\item loglik log likelihood
}
}
\description{
This is primarily for plugging in final A (e.g., from PEM) and redo EM to calculate log lik
for EBIC calculations. Response probabilities estimated within the algorithm
with A and Q fixed.
}
