% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{get_em_classify}
\alias{get_em_classify}
\title{EM algorithm for estimating c, g and mixing parameter pi}
\usage{
get_em_classify(X, Q, A_in, err_prob, model = "DINA")
}
\arguments{
\item{X}{N by J data matrix}

\item{Q}{J by K structural matrix}

\item{A_in}{C by K latent attributes}

\item{err_prob}{error probabilities 0.2}

\item{model}{"DINA" (default) or "DINO"}
}
\value{
\itemize{
\item nu
\item c
\item g
\item loglik
\item Z_shrink
}
}
\description{
EM algorithm for estimating c, g and mixing parameter pi
}
