% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_I}
\alias{get_I}
\title{get none, singleton, doubleton, ... until full binary patterns}
\usage{
get_I(D, M)
}
\arguments{
\item{D}{sum of binary codes (if zero, forced to be 1)}

\item{M}{dimension of binary codes}
}
\value{
a binary matrix of size (M + M choose 2 + ...+ M choose D)by M
}
\description{
Within each k-ton, we order from the largest to smallest
}
\examples{
get_I(3,3)
get_I(2,3)
get_I(1,3)
get_I(0,3)
}
