% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_ideal_resp}
\alias{get_ideal_resp}
\title{get ideal response for DINA}
\usage{
get_ideal_resp(Q, Z)
}
\arguments{
\item{Q}{Q matrix of size J by K}

\item{Z}{latent attribute profiles of individuals; N by K}
}
\value{
binary ideal response matrix of size N by J
}
\description{
replace Z by 1-Z and 1- the resultant matrix gives the
ideal response matrix for DINO
}
\examples{
Z <- rbind(c(1,1,1),c(0,0,1))
Q <- matrix(c(0,0,0, 0,0,1, 0,1,0,
              0,1,1, 1,0,0, 1,0,1,
              1,1,0, 1,1,1),ncol=3,byrow=TRUE)
get_ideal_resp(Q,Z)

get_ideal_resp(Q=matrix(c(1,1,0,
                          0,0,1,
                          0,1,0,
                          1,0,1),nrow=4,byrow=TRUE),
               Z = rbind(c(1,1,1),c(0,0,0)))
}
