% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{get_initial_s}
\alias{get_initial_s}
\title{get initial Q and Z}
\usage{
get_initial_s(N, Q0, Z0)
}
\arguments{
\item{N}{sample size}

\item{Q0}{J by K initial Q}

\item{Z0}{N by K initial Z}
}
\value{
\itemize{
\item Q_ini
\item Z_ini
}
}
\description{
get initial Q and Z
}
