% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{get_PEM}
\alias{get_PEM}
\title{PEM FOR SHRINKAGE - this is to take a candidate latent attribute set}
\usage{
get_PEM(X, Q, A, lambda, c_ini, g_ini, nu_ini, thres_c, model = "DINA")
}
\arguments{
\item{X}{N by J data matrix}

\item{Q}{J by K structural matrix}

\item{A}{K-column latent attributes; candidates to be selected!}

\item{lambda}{negative penalty parameter}

\item{c_ini}{initial 1-slipping parameters}

\item{g_ini}{initial guessing parameters}

\item{nu_ini}{initial potentials}

\item{thres_c}{threshold for making sure M step does not generate negative
probabilities (algorithm 1 of Xu and Gu JMLR 2019)}

\item{model}{"DINA" (default) or "DINO"}
}
\value{
\itemize{
\item nu vector of potential ("Delta")
\item c 1-slipping parameter
\item g guessing parameter
}
}
\description{
PEM FOR SHRINKAGE - this is to take a candidate latent attribute set
}
