% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_rp}
\alias{get_rp}
\title{calculates marginal probabilities of all the 2^J possible
response patterns}
\usage{
get_rp(A, Q, c, g, I_full)
}
\arguments{
\item{A}{K-column, latent attribute profile matrix}

\item{Q}{M by K Q matrix}

\item{c}{M by 1: 1-slipping}

\item{g}{M by 1 guessing parameter}

\item{I_full}{J-column matrix; Each row is one of 2^J patterns under consideration
can be generated by \code{rbind(rep(0,J),get_I(J,J))}}
}
\description{
calculates marginal probabilities of all the 2^J possible
response patterns
}
\examples{
A <- rbind(c(1,1,1),c(0,0,1))
Q <- rbind(c(1,0,0),c(1,0,0),c(0,1,0),c(1,1,1))
c <- c(0.8,0.8,0.8,0.8)
g <- 1-c(0.8,0.8,0.8,0.8)
I_full <- rbind(rep(0,4), get_I(4,4))
get_rp(A,Q,c,g,I_full)
}
