% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nchoosek_prac}
\alias{nchoosek_prac}
\title{n choose k - dealing with large values.}
\usage{
nchoosek_prac(n, k)
}
\arguments{
\item{n, k}{integers (n>=k)}
}
\description{
n choose k - dealing with large values.
}
