% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{perform_shrink}
\alias{perform_shrink}
\title{performs shrinkage estimation and use Extended BIC (EBIC) to select the
best tuning parameter in the penalized likelihood from an entire solution path}
\usage{
perform_shrink(
  X,
  Q_est,
  Z_candi,
  lambda_vec,
  c_ini,
  g_ini,
  is_sub = 0,
  model = "DINA"
)
}
\arguments{
\item{X}{N by J data matrix}

\item{Q_est}{J by K Q after screening by \code{\link{adg_em}} (one-level version)}

\item{Z_candi}{K-column candidate latent attributes (candidates;
\code{\link{adg_em}} (one-level version) )}

\item{lambda_vec}{a decreasing sequence of negative penalty parameters}

\item{c_ini}{J initial values of 1-slipping parameters}

\item{g_ini}{J initial values of guessing parameters}

\item{is_sub}{use subsample (1) or not (0; default) - NB: not implemented.}

\item{model}{"DINA" (default) or "DINO"}
}
\value{
\itemize{
\item A_final final estimated set of latent attribute patterns that made the cut
\item EBIC_vec vector of EBIC values
\item size_vec for each tunning parameter, the size of latent attribute patterns
that made the cut.
}
}
\description{
performs shrinkage estimation and use Extended BIC (EBIC) to select the
best tuning parameter in the penalized likelihood from an entire solution path
}
