% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct.R
\name{reconstruct}
\alias{reconstruct}
\title{reconstructs N by J data matrix using the estimated item parameters c, g
and other quantities}
\usage{
reconstruct(Z_est, Q_est, c, g, model = "DINA")
}
\arguments{
\item{Z_est}{N by K of estimated attributes}

\item{Q_est}{J by K of estimated Q}

\item{c}{J vector of 1-slipping parameters}

\item{g}{J vector of guessing parameters}

\item{model}{"DINA" (default), or "DINO"}
}
\value{
N by J matrix; use p_correct >0.5 as reconstruction
}
\description{
To reconstruct without c or g, use \code{get_ideal_resp(Q_est,Z_est)}
}
\examples{
set.seed(60)
Z_est <- matrix((rnorm(200)>0)+0,nrow=50,ncol=4)
Q_est <- matrix(c(1,1,1,1,0,0,1,0,1,0,1,0,1,1,1,0,0,0,0,1,1,0,0,1),ncol=4,
byrow=TRUE)
c <- rep(0.8,nrow(Q_est))
g <- 1-c
reconstruct(Z_est,Q_est,c,g)
}
\seealso{
\code{\link{get_ideal_resp}}
}
