% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sort_binmat}
\alias{sort_binmat}
\title{order binary matrix by row}
\usage{
sort_binmat(mat, ...)
}
\arguments{
\item{mat}{binary matrix}

\item{...}{arguments in \code{order}}
}
\value{
binary matrix (ordered by row)
}
\description{
order binary matrix by row
}
\examples{
sort_binmat(matrix(c(1, 1, 1, 0 , 1, 0, 1, 1,0),nrow=3,byrow=TRUE))
}
