% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unique_sort_binmat}
\alias{unique_sort_binmat}
\title{unique and ordered binary matrix by row}
\usage{
unique_sort_binmat(mat, ...)
}
\arguments{
\item{mat}{binary matrix}

\item{...}{arguments in \code{order}}
}
\value{
binary matrix (ordered by row)
}
\description{
unique and ordered binary matrix by row
}
\examples{
unique_sort_binmat(matrix(c(1, 1, 1, 1,1,1, 0 , 1, 0, 1, 1,0),nrow=4,byrow=TRUE))
}
