% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{binary}
\alias{binary}
\title{convert decimal integers to binary codes}
\usage{
binary(x, k = NULL)
}
\arguments{
\item{x}{a vector of base 10 integers}

\item{k}{length of binary codes (if no specified, this will be the
length of the binary representation for the maximum in \code{x}).
Default to \code{NULL}.}
}
\value{
A matrix of binary forms of x (by row)
}
\description{
convert decimal integers to binary codes
}
\examples{
decimals <- c(10,3,5)
binary(decimals)
binary(decimals,6)
}
